move ServiceMsgClientConn from util to client package
