+++ client/msg_service.go
@@ -1,4 +1,4 @@
-package util
+package client
 
 import (
 	"context"
